backend/requirements.txt
fastapi
uvicorn[standard]
pydantic==2.11.7
pydantic[email]
python-dotenv
pyyaml
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
psycopg2-binary
pymongo
redis
google-generativeai
openai
groq
motor
httpx
click
h11
starlette
typing_extensions
typing_inspection
annotated_types
anyio
sniffio
ecdsa
six
python-multipart
pymysql
elasticsearch
google-cloud
google-cloud-storage
google-cloud-pubsub
google-cloud-bigquery
distro
httpcore
requests
elastic_transport
email-validator
google-resumable-media
google-api-core
google-cloud-core
python-dateutil
packaging
mcp
asyncpg
arq
asyncpg
arq
arq
tenacity
